requirements.txt
fastapi
uvicorn
pyserial
opencv-python
websockets
pyaudio